@@ -1,61 +1,82 @@
 {
-  "name": "matthew-portfolio",
-  "version": "0.1.0",
-  "private": true,
-  "proxy": "http://localhost:8081",
-  "engines": {
-    "node": "7.10.1",
-    "npm": "4.1.2"
-  },
-  "devDependencies": {
-    "babel-watch": "^2.0.7",
-    "concurrently": "^3.5.0",
-    "husky": "^0.14.3",
-    "lint-staged": "^4.0.2",
-    "prettier": "^1.5.3"
-  },
-  "dependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "compression": "^1.7.0",
-    "express": "^4.15.3",
-    "node-sass": "^4.9.4",
-    "path": "^0.12.7",
-    "react": "^16.4.2",
-    "react-dom": "16.4.2",
-    "react-modal": "^2.2.4",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "2.0.4",
-    "react-scroll": "^1.5.4",
-    "request": "^2.79.0",
-    "sw-precache": "^5.2.0"
-  },
-  "scripts": {
-    "start": "concurrently \"PORT=8080 react-scripts start\" \"PORT=8081 babel-watch index.js\"",
-    "precommit": "lint-staged",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "deploy": "now && now alias",
-    "now-dev": "react-scripts start",
-    "now-build": "react-scripts build"
-  },
-  "lint-staged": {
-    "src/**/*.js": [
-      "prettier --trailing-comma es5 --single-quote --write",
-      "git add"
-    ],
-    "index.js": [
-      "prettier --trailing-comma es5 --single-quote --write",
-      "git add"
-    ]
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+	"name": "matthew-portfolio",
+	"version": "0.1.0",
+	"private": true,
+	"proxy": "http://localhost:8081",
+	"engines": {
+		"node": "12.0.0",
+		"npm": "6.13.4"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.10.2",
+		"babel-core": "^6.26.3",
+		"babel-preset-latest": "^6.24.1",
+		"babel-watch": "^2.0.7",
+		"concurrently": "^3.5.0",
+		"eslint": "^6.6.0",
+		"eslint-config-airbnb": "^18.2.0",
+		"eslint-config-prettier": "^6.11.0",
+		"eslint-plugin-jsx-a11y": "^6.3.1",
+		"eslint-plugin-prettier": "^3.1.4",
+		"eslint-plugin-react": "^7.20.0",
+		"husky": "^4.2.3",
+		"lint-staged": "^10.0.8",
+		"prettier": "^1.19.1",
+		"pretty-quick": "^2.0.1",
+		"stylelint": "^13.6.1",
+		"stylelint-order": "^4.1.0",
+		"typescript": "^3.9.5"
+	},
+	"dependencies": {
+		"@carbon/icons-react": "^10.13.0",
+		"compression": "^1.7.0",
+		"express": "^4.15.3",
+		"node-sass": "^4.14.1",
+		"path": "^0.12.7",
+		"prop-types": "^15.7.2",
+		"react": "^16.13.1",
+		"react-dom": "16.13.1",
+		"react-router": "^5.1.2",
+		"react-router-dom": "^5.1.2",
+		"react-scripts": "3.4.1",
+		"react-scroll": "^1.5.4",
+		"react-scroll-parallax": "^2.3.3",
+		"react-spring": "^9.0.0-rc.3",
+		"request": "^2.79.0",
+		"styled-components": "^5.1.1",
+		"sw-precache": "^5.2.0"
+	},
+	"scripts": {
+		"start": "concurrently \"PORT=8080 react-scripts start\" \"PORT=8081 babel-watch index.js\"",
+		"build": "react-scripts build",
+		"test": "react-scripts test --env=jsdom",
+		"deploy": "now && now --target production",
+		"now-dev": "react-scripts start",
+		"now-build": "react-scripts build",
+		"lint": "eslint .",
+		"lint-fix": "eslint --fix .",
+		"stylelint": "stylelint src",
+		"stylelint-fix": "stylelint --fix src"
+	},
+	"husky": {
+		"hooks": {
+			"pre-commit": "lint-staged"
+		}
+	},
+	"lint-staged": {
+		"**.{js,jsx}": [
+			"pretty-quick --staged",
+			"eslint --fix",
+			"git add"
+		],
+		"**.{scss,css}": [
+			"stylelint --fix"
+		]
+	},
+	"browserslist": [
+		">0.2%",
+		"not dead",
+		"not ie <= 11",
+		"not op_mini all"
+	]
 }
