@@ -1,16 +1,12 @@
-import { Component } from 'react';
-import { withRouter } from 'react-router';
+import { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 
-class ScrollToTop extends Component {
-  componentDidUpdate(prevProps) {
-    if (this.props.location !== prevProps.location) {
-      window.scrollTo(0, 0);
-    }
-  }
+const ScrollToTop = ({ children }) => {
+	const location = useLocation();
+	useEffect(() => {
+		window.scrollTo(0, 0);
+	}, [location]);
+	return children;
+};
 
-  render() {
-    return this.props.children;
-  }
-}
-
-export default withRouter(ScrollToTop);
+export default ScrollToTop;
