@@ -1,29 +1,40 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 
-//styles
 import './navLink.scss';
 
-class NavLink extends Component {
-  render() {
-    return (
-      <li className="nav-link-item nav-link__space">
-        <a
-          href={this.props.url}
-          className={`nav-link mg--bold ${this.props.className}`}
-          onClick={this.props.openModal}
-        >
-          <span>{this.props.name}</span>
-        </a>
-      </li>
-    );
-  }
-}
+const NavLink = ({ url, externalLink, className, openModal, name }) => {
+	return (
+		<li className="nav-link-item nav-link__space">
+			{externalLink ? (
+				<a href={url}
+className={`nav-link mg--bold ${className || ''}`}
+onClick={openModal}>
+					<span>{name}</span>
+				</a>
+			) : (
+				<Link to={url}
+className={`nav-link mg--bold ${className || ''}`}>
+					<span>{name}</span>
+				</Link>
+			)}
+		</li>
+	);
+};
+
+NavLink.defaultProps = {
+	className: '',
+	externalLink: false,
+	openModal: null,
+};
 
 NavLink.propTypes = {
-  className: PropTypes.string,
-  name: PropTypes.string.isRequired,
-  url: PropTypes.string,
+	className: PropTypes.string,
+	name: PropTypes.string.isRequired,
+	url: PropTypes.string.isRequired,
+	externalLink: PropTypes.bool,
+	openModal: PropTypes.func,
 };
 
 export default NavLink;
