@@ -1,6 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Routes from './Routes';
+import { ParallaxProvider } from 'react-scroll-parallax';
+import App from './App';
 import './index.scss';
 
-ReactDOM.render(<Routes />, document.getElementById('root'));
+ReactDOM.render(
+	<ParallaxProvider>
+		<App />
+	</ParallaxProvider>,
+	document.getElementById('root')
+);
